refactor: PaymentClient OCP 원칙 적용
