+++ src/main/resources/application.properties
@@ -4,8 +4,8 @@ spring.sql.init.platform=prod
 security.jwt.token.secret-key=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno
 security.jwt.token.expire-length=360000
 
-security.payment.secret-key=test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
-security.payment.url=https://api.tosspayments.com
+security.toss.payment.url=https://api.tosspayments.com
+security.toss.payment.secret-key=test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
 
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
